Implemented DebugBar instead of Statistics.cs, which now exposes the IStatistics service and draws a pretty background.
