+++ src/Engine/Debugging/Graphs/Implementations/ChunkGraphs/LightenQ.cs
@@ -61,7 +61,7 @@ namespace VoxeliqEngine.Debugging.Graphs.Implementations.ChunkGraphs
 
         public override void DrawGraph(GameTime gameTime)
         {
-            BasicShapes.DrawSolidPolygon(this.PrimitiveBatch, Background, 4, Color.Black, true);
+            BasicShapes.DrawSolidPolygon(this.PrimitiveBatch, BackgroundPolygon, 4, Color.Black, true);
 
             float x = Bounds.X;
             float deltaX = Bounds.Width / (float)ValuesToGraph;
