+++ src/Engine/VoxeliqEngine.Windows.MonoGame.GL.csproj
@@ -139,6 +139,7 @@
     <Compile Include="Debugging\Console\KeyEventArgs.cs" />
     <Compile Include="Debugging\Console\OutputLine.cs" />
     <Compile Include="Debugging\Console\Renderer.cs" />
+    <Compile Include="Debugging\DebugBar.cs" />
     <Compile Include="Debugging\Graphs\DebugGraph.cs" />
     <Compile Include="Debugging\Graphs\GraphManager.cs" />
     <Compile Include="Debugging\Graphs\Implementations\ChunkGraphs\BuildQ.cs" />
@@ -151,7 +152,6 @@
     <Compile Include="Debugging\Ingame\BoundingBoxRenderer.cs" />
     <Compile Include="Debugging\Ingame\InGameDebugger.cs" />
     <Compile Include="Debugging\Profiling\Profiler.cs" />
-    <Compile Include="Debugging\Statistics.cs" />
     <Compile Include="Debugging\Timing\Layout.cs" />
     <Compile Include="Debugging\Timing\TimeRuler.cs" />
     <Compile Include="Graphics\Camera.cs" />
